Version 0.18.1
E83 - bug correction: In function SQLite2df. When the 2 different sensor models were installed on the same channel of an ASE box the automatic recognition of sensor model was producing a crash of the App with the new wide shape of airsenseur.db. Now corrected. The names of sensors is the last one on the channel.